Rename calculate.py as calculator.py
