+++ taxcalc/__init__.py
@@ -1,7 +1,7 @@
 """
 Specify what is available to import from the taxcalc package.
 """
-from taxcalc.calculate import *
+from taxcalc.calculator import *
 from taxcalc.policy import *
 from taxcalc.behavior import *
 from taxcalc.consumption import *
